Fixed various bugs with baking cookies and accessing the cooky store.
Made it so alert messages Don't use dialogs to notify the player of events while they are in baking mode without having to interupt it.
Made it so the store open and close sounds can be skipped by pressing the enter key to go in and out of the area faster.
Fixed a bug where the baking timer was still in the active state even though you didn't have anymore autocookies in your inventory.