Fixed some bugs, and added some more features.
Added store anbience volume and pitch to the audio settings menu.
Brought back the information that was given to the user after purchasing an item. It will now tell you the next upgrade cost of all items that are bought in the cooky store.
Added some more music to play with.