Replaced manual baking sounds, added doors, and added menu music to most menus of the game.
Added menu music to sertain menus in the game. You can adjust the volume of the music by pressing the page up and page down keys. The shop menus will still have it's own store music.
Added a door that will randomly spawn on the map while you're on the boss area in the toy mania mini game. You can't open this door unless you've collected all of the keys that are required to unlock it.
Replaced the manual baking sounds with new ones.