Minor game enhancements.
Replaced the boss hitting sounds with new ones in the toy mania mini game.
Decreased the amount of keys you need to collect to unlock the door in the toy mania mini game. This value will also be spoken when trying to unlock the door.
Made it so the ammo of the machine gun resets upon redrawing it in the toy mania mini game.