You can no longer have all positive or all Negative auto and manual baking events set in the game settings menu. You are now required to select at least one positive and one Negative event in each category before you are allowed to save the game settings.
Made it so you are required to at least select 1 Positive and one negative auto baking event out of the remaining auto baking events in the auto baking events list before saving the game settings in the game settings menu.
Made it so you are required to at least select 1 Positive and one negative manual baking event out of the remaining manual baking events in the manual baking events list before saving the game settings in the game settings menu.